Split model.go to more files
